requirements.txt
mcp>=1.0.0
fastmcp>=0.1.0
requests>=2.31.0
flask>=3.0.0
pycryptodome>=3.20.0
PyJWT>=2.8.0
